src/Makefile: Fix GCC version detection
It seems newer versions of GCC tend to now be compiled with
--with-gcc-major-version-only which means -dumpversion only prints the
major version number and we end up with the same number for each of
major/minor/sub.

For example, we might think we are using GCC 11.11.11 when in fact it is
actually 11.2.0

However GCC 7 introduced a new option -dumpfullversion which always
prints the three parts.

Furthermore we can take advantage of the fact that GCC will stop
processing after hitting the first successful -dump* option.

So if we put -dumpfullversion before -dumpversion then this will work on
both new and old versions of GCC.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>