Use an ac_btree_t to store the file entries
Each dentry (directory entry) has a jfiles member which contains a list
of all the files/directories contained in that directory.

Previously this was done with an array, where we'd loop through this
array and construct a full file path for each entry to compare against
the path we are looking for.

While we're not dealing with directories with *huge* numbers of
entries, it would be more efficient to use something like a binary tree
to store these entries.

So that's what this does, convertes the dentry->jfiles from an array to
an ac_btree_t.

It also means we no longer need to construct the full path when doing
lookups, as we simply lookup the component entry we are looking for.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>