Try to show the proper st_nlink for directories
This value (on XFS and ext4 at least) seems to be the number of
directories contained within plus '.' & '..'

Try to replicate this here, a lot of these values will be filled in
lazily and show 0 by default until the directory has been accessed and
information pulled from jamendo.

We track this st_nlink value via a nlink member of the jf_file
structure.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>